TODO/requirements.txt
certifi==2017.11.5
chardet==3.0.4
Django==1.9
django-crispy-forms==1.6.0
django-markdown-deux==1.0.5
django-pagedown==0.1.1
djangorestframework==3.3.3
djangorestframework-jwt==1.8.0
idna==2.6
markdown2==2.3.1
Pillow==3.1.0
pkg-resources==0.0.0
PyJWT==1.4.0
requests==2.18.4
urllib3==1.22
